Medical Expenditure Panel Survey (MEPS) Data

The MEPS dataset contains large-scale survey data from the United States, focusing on health services usage, costs, and insurance coverage. This dataset is restricted to individuals aged 21 to 64 years. It includes outpatient cost data with some zero expenditure values for model adjustment.

Usage

MEPS2001

Format

A data frame with 3328 observations on the following variables:

educ: Education status (numeric)

age: Age (numeric)

income: Income (numeric)

female: Gender (binary)

vgood: Self-reported health status, very good (numeric)

good: Self-reported health status, good (numeric)

hospexp: Hospital expenditures (numeric)

totchr: Total number of chronic diseases (numeric)

ffs: Family support (numeric)

dhospexp: Dummy variable for hospital expenditures (binary)

age2: Age squared (numeric)

agefem: Interaction between age and gender (numeric)

fairpoor: Self-reported health status, fair or poor (numeric)

year01: Year of survey (numeric)

instype: Type of insurance (numeric)

ambexp: Ambulatory expenditures (numeric)

lambexp: Log of ambulatory expenditures (numeric)

blhisp: Ethnicity (binary)

instype_s1: Insurance type, version 1 (numeric)

dambexp: Dummy variable for ambulatory expenditures (binary)

lnambx: Log-transformed ambulatory expenditures (numeric)

ins: Insurance status (binary)

Source

2001 Medical Expenditure Panel Survey by the Agency for Healthcare Research and Quality.

Examples

data(MEPS2001)
selectEq  <- dambexp ~ age + female + educ + blhisp + totchr + ins + income
outcomeEq <- lnambx ~ age + female + educ + blhisp + totchr + ins
dispersion  <- ~ age + female + totchr + ins
correlation  <- ~ age
fit <- heckmanGE(selection = selectEq,
                outcome = outcomeEq,
                dispersion = dispersion,
                correlation = correlation,
                data = MEPS2001)
#> Error in eval(mfS): object 'selectEq' not found
summary(fit)
#> Error: object 'fit' not found