meatCL.heckmanGE Compute the Meat Matrix for a Heckman-Ge Model with Clustering

This function calculates the meat matrix for a Heckman-Ge model, which is used in the context of clustered standard errors. The meat matrix represents the variability of the estimated parameters and is a crucial component for robust inference.

Usage

meatCL.heckmanGE(
  x,
  cluster = NULL,
  type = NULL,
  cadjust = TRUE,
  multi0 = FALSE,
  ...
)

Arguments

x

An object of class heckmanGE containing the results from a Heckman-Ge model fit.

cluster

A vector or a data frame specifying the cluster variable(s). If NULL, the default clustering variable is used.

type

The type of heteroscedasticity-consistent (HC) estimator to use. Options are "HC0", "HC1", "HC2", or "HC3". Defaults to "HC0".

cadjust

A logical value indicating whether to adjust for the number of clusters. Defaults to TRUE.

multi0

A logical value indicating whether to include a column of ones in the cluster variable matrix. Defaults to FALSE.

...

Additional arguments passed to other methods.

Value

A matrix representing the meat component of the robust covariance matrix estimator for the Heckman-Ge model.