PNAD Continua de 2024, 2 trimestre

The Continuous National Household Sample Survey (PNAD Continua) for the second quarter of 2024 is an important source of statistical data in Brazil, conducted by the Brazilian Institute of Geography and Statistics (IBGE). The survey aims to provide up-to-date information on the socioeconomic characteristics of the Brazilian population, covering topics such as employment, income, education, and other crucial aspects for the formulation of public policies and economic and social studies.

Usage

pnadC_y2024q2

Format

A data frame with 326018 observations on the following variables:

PSU: Primary Sampling Unit identifier (factor)

weight: Survey weight (numeric)

age: Age of the respondent (numeric)

participation: Labor force participation status (factor)

male: Male indicator (binary)

white: White indicator (binary)

hhold_head: Household head indicator (binary)

hhold_spouse: Spouse of household head indicator (binary)

yearsSchooling: Total years of schooling completed (numeric)

classWorker_employer: Employer indicator (binary)

classWorker_selfEmployed: Self-employed indicator (binary)

ln_salary: Natural logarithm of salary (numeric)

Examples

data(pnadC_y2024q2)
pnadC_y2024q2 <- pnadC_y2024q2[1:10000,]
attach(pnadC_y2024q2)
#> The following object is masked from MEPS2001:
#> 
#>     age
selectEq  <- participation ~ age + I(age^2) +
 male + white + yearsSchooling +
 hhold_head + hhold_spouse
outcomeEq <- ln_salary ~ age + I(age^2) +
 male + white + yearsSchooling +
 classWorker_employer + classWorker_selfEmployed
outcomeD  <- ~ age + I(age^2) +
 male + white + yearsSchooling +
 classWorker_employer + classWorker_selfEmployed
outcomeC  <- ~ male + yearsSchooling
fit_heckmanGE <- heckmanGE(selection   = selectEq,
                           outcome     = outcomeEq,
                           dispersion  = outcomeD,
                           correlation = outcomeC,
                           data = pnadC_y2024q2,
                           weights = weight,
                           cluster = ~PSU)
#> Error in eval(mfS): object 'selectEq' not found
                           summary(fit_heckmanGE)
#> Error in eval(expr, envir, enclos): object 'fit_heckmanGE' not found