Simulation dataset for the heckmanGE example

This dataset contains simulated data used to illustrate the functionality of the heckmanGE model. The data includes variables used in selection, outcome, dispersion, and correlation equations.

Usage

simulation

Format

A data frame with 10,000 observations on the following variables:

y_o: Outcome variable from the simulated model (numeric)

y_s: Selection indicator, 1 if selected, 0 otherwise (binary)

prob_s: Probability of selection (numeric)

x1: Simulated predictor from a normal distribution (numeric)

x2: Simulated predictor from a Poisson distribution (numeric)

x3: Simulated binary predictor (binary)

x4: Simulated predictor from a normal distribution with mean 2 and sd 2 (numeric)

x5: Simulated predictor from a Poisson distribution with lambda 1.5 (numeric)

Examples

data(simulation)
selectEq  <- y_s ~ x1 + x2 + x4
outcomeEq <- y_o ~ x1 + x2 + x3
outcomeD  <- ~ x1 + x5
outcomeC  <- ~ x3 + x4
fit_heckmanGE <- heckmanGE(selection   = selectEq,
                           outcome     = outcomeEq,
                           dispersion  = outcomeD,
                           correlation = outcomeC,
                           data = simulation)
#> Error in eval(mfS): object 'selectEq' not found
summary(fit_heckmanGE)
#> Error in eval(expr, envir, enclos): object 'fit_heckmanGE' not found